requirements.txt
streamlit==1.40.0
pandas>=2.0.0
streamlit-option-menu>=0.3.0
apify-client>=1.0.0
python-dotenv>=1.0.0
